-added more constraints to stepper in addession form -fixed issues with verifypage in addession -added fields to support sleep + anxiety tracking -added anxiety and sleep tracking fields -removed console.logs from app -made addession folder and moved relevant components into it
